Added new functions to estimate initial infected and deaths
I have taken data from the SEIR simulation of https://covid19-scenarios.org/ and estracted data for death and severity percentage so to make an estimate of the unknow real infected population in my model, and then estimates deaths and severe cases from my model. 

There is still a long way to go, but it's an improvement.