Added new estimates for peak scenarios and analysis of the doubling time
I have added a new estimate for the peak scenarios that is based on computing quantiles/averages of the peak intensities and position rather than by computing the peak position and intensity of the mean number of infected people. This should give more robust information.

Furthermore, I hav added a few checks about the exponential growth in Okinawa, and it seems that my model is way off. It estimates a doubling time between 6 and 12 days, whereas, the real data predict a doubling time of about 5 days. This suggests that I should use the infection time, the incubation time and the doubling time as free parameters and fit R0 from these three values.