Added new estimates of R0 from okinawa doubling time
Now the simulation uses an estimation of R0 that predicts a doubling time of 5 days, as in okinawa. Former estimated of R0 from Ferretti, indeed predicted a doubling time that was too large than what is currently happening now in the island. The bad news is that R0 is larger than 3.