Implemented crude code to compute R0 dynamics from Okinawa data
I have implemented all the functions that I need to compute R0 dynamics from the measured data in Okinawa. The code is a frankenstein patch of a job, but it should give the right outputs. I now need to polish it a bit, and run an optimization with a little bit more of points to start.