Updated SEIR_MWE to compute R0 suppression with minimization
I have updated the SEIR MWE so that it now computes R0 suppression by minimizing the deviation between the cumulative infected from a SEIR simulation and from a known time series. for the moment the time series is a SEIR simulation too with some extra noise on top.

The optimization finds the correct parameters for the known SEIR time series in a stable way. This gives a reasonable outline of how I should clean the code and make it work for the Okinawa data.