Corrected error and stress test of the MWE
I have corrected an error on how I computed R0 suppression and then I performed some stress test to see wether the code works fine when the time series of the infected is very short and noisy. 

It seems that the optimization returns the correct R0, R0 suprression and time of suppression, but fails miserably to address the true number of initial infected and initial exposed.